src/pyproject.toml
[project]
name = "payout-service"
version = "0.1.0"
description = "Test task for SmartCollect"
authors = [
    {name = "S3raphimCS",email = "700520serj@mail.ru"}
]
readme = "../README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (<6)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-filter (>=25.2,<26.0)"
]

[tool.poetry]
packages = [{include = "server"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "server.settings"
pythonpath = ["server", "."]
testpaths = ["server/tests"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--nomigrations --reuse-db"

[tool.ruff]
line-length = 88
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C90"]
ignore = ["E501", "E203"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["D", "S"]
"manage.py" = ["D"]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["apps", "server"]